feat(cli): add train command (#1023)
<!-- markdownlint-disable -->

Sub-PR 6 of 6 for the [CLI RFC
implementation (umbrella: #1024).

Supersedes #940 (old sequential topology).

- Create `speculators.train.cli` with the training entrypoint (moved
from `scripts/train.py`)
- Add `speculators.train.__main__` for `torchrun -m speculators.train`
- Wire a thin typer dispatcher as `speculators train` under the Pipeline
help panel (forwards all args to `TrainConfig.resolve`)
- Fix `speculators train --help` to pass through to the underlying
argparse parser (shows all training flags instead of empty typer help)
- Original `scripts/train.py` becomes a deprecation shim (will be
removed in v0.9.0)
- Update docs (`docs/user_guide/tutorials/train.md`,
`docs/cli/index.md`) with new CLI references and fix broken links
- Update `scripts/benchmark.py` imports to use new module path; remove
stale `sys.path.insert`
- Update test imports (`test_draft_config_init.py`,
`test_rope_config.py`) to use `speculators.train.cli`
- Remove stale `legacy_data` parameter from `create_train_val_loaders`
call (removed in c80a242)
- Add 2 smoke tests for the train command

Depends on: #1018 (CLI infrastructure)

```bash
git diff main:scripts/train.py cli/train:src/speculators/train/cli.py
```

```
$ pytest tests/unit/cli/ -v
tests/unit/cli/test_cli.py::TestRootApp::test_no_args_shows_help PASSED
tests/unit/cli/test_cli.py::TestRootApp::test_help PASSED
tests/unit/cli/test_cli.py::TestRootApp::test_version PASSED
tests/unit/cli/test_cli.py::TestRootApp::test_pipeline_commands_in_help PASSED
tests/unit/cli/test_cli.py::TestRootApp::test_tools_commands_in_help PASSED
tests/unit/cli/test_cli.py::TestConvertCommand::test_help PASSED
tests/unit/cli/test_cli.py::TestConvertCommand::test_algorithm_choices_in_help PASSED
tests/unit/cli/test_cli.py::TestConvertCommand::test_missing_required_args PASSED
tests/unit/cli/test_cli.py::TestTrainCommand::test_help PASSED
tests/unit/cli/test_cli.py::TestTrainCommand::test_train_appears_in_pipeline_panel PASSED
======================== 10 passed in 0.28s ========================
```

I have filled in:

- [x] The purpose of the PR, such as "Fix some issue (link existing
issues this PR will resolve)".
- [x] The test plan/results, such as providing test command and pasting
the results.
- [x] (Optional) The necessary documentation update.
- [x] I (a human) have written or reviewed the code in this pr to the
best of my ability.

---------
