+++ src/speculators/cli/__init__.py
@@ -13,6 +13,7 @@ import typer
 from speculators.cli.convert import convert
 from speculators.cli.prepare_data import prepare_data
 from speculators.cli.stitch import stitch_command
+from speculators.cli.train import train_command
 
 __all__ = ["app"]
 
@@ -42,4 +43,14 @@ def _main(
 
 app.command(rich_help_panel="Pipeline")(prepare_data)
 app.command(name="stitch-mtp", rich_help_panel="Pipeline")(stitch_command)
+app.command(
+    name="train",
+    rich_help_panel="Pipeline",
+    context_settings={
+        "allow_extra_args": True,
+        "allow_interspersed_args": False,
+        "ignore_unknown_options": True,
+        "help_option_names": [],
+    },
+)(train_command)
 app.command(rich_help_panel="Tools")(convert)
