+++ tests/unit/cli/test_cli.py
@@ -28,6 +28,7 @@ class TestRootApp:
         assert result.exit_code == 0
         assert "prepare-data" in result.output
         assert "stitch-mtp" in result.output
+        assert "train" in result.output
 
     def test_tools_commands_in_help(self):
         result = runner.invoke(app, ["--help"])
@@ -92,3 +93,17 @@ class TestStitchCommand:
     def test_missing_required_args(self):
         result = runner.invoke(app, ["stitch-mtp"])
         assert result.exit_code != 0
+
+
+class TestTrainCommand:
+    def test_help(self):
+        result = runner.invoke(app, ["train", "--help"])
+        assert result.exit_code == 0
+        assert "--verifier-name-or-path" in result.output
+        assert "--config" in result.output
+        assert "--speculator-type" in result.output
+
+    def test_train_appears_in_pipeline_panel(self):
+        result = runner.invoke(app, ["--help"])
+        assert result.exit_code == 0
+        assert "train" in result.output
